:recycle:[refactor]: Report Section 분리
